+++ src/main/java/com/github/danielwegener/logback/kafka/KafkaAppender.java
@@ -112,14 +112,23 @@ public class KafkaAppender<E> extends KafkaAppenderConfig<E> {
 
     @Override
     protected void append(E e) {
-        final byte[] payload = encoder.doEncode(e);
+        final byte[] payload = encoder.encode(e);
         final byte[] key = keyingStrategy.createKey(e);
-        final ProducerRecord<byte[], byte[]> record = new ProducerRecord<byte[],byte[]>(topic, key, payload);
+
+        final Long timestamp;
+        if (e instanceof ILoggingEvent) {
+            timestamp = ((ILoggingEvent) e).getTimeStamp();
+        } else {
+            timestamp = null;
+        }
+        final ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(topic, null, timestamp, key,
+                payload);
+
         deliveryStrategy.send(lazyProducer.get(), record, e, failedDeliveryCallback);
     }
 
     protected Producer<byte[], byte[]> createProducer() {
-        return new KafkaProducer<byte[], byte[]>(new HashMap<String, Object>(producerConfig));
+        return new KafkaProducer<>(new HashMap<>(producerConfig));
     }
 
     private void deferAppend(E event) {